Add logic for both connecting and disconnecting from Airstory when the user provides a token.
When a user enters their token, Airstory\Settings\save_profile_settings() fires the airstory_user_connect action; Airstory\Connection\register_connection() listens for this as a signal for it to grab the user profile (details about their account on Airstory), then registers a new "target" within Airstory.

The reason for getting the profile is two-fold:
1. To be able to say "You're currently authenticated as this user" below the "disconnect from Airstory" button
2. The API route for targets is /users/{email}/targets, where {email} is the email account tied to the user's Airstory account. Considering this will not always be the same as the user's WordPress email, the profile ensures we can maintain this connection.

Two items are stored in user meta: the (reduced) profile information (with the important item being the email address) and the UUID of the current target.

The remove_connection() function, on the other hand, deletes the connection and corresponding user meta, forcing them to re-authenticate.