Add either .dependency-met or .dependency-unmet to each table row in the Airstory compatibility checks.
As we're only using three custom styles, I chose to go the route of inline CSS (rather than introducing an external stylesheet that would *only* be loaded on this one page anyway. If more styles are added down the road, we may want to revisit this and break it off into a separate asset.