If Airstory\Connection\register_connection() fails, explicitly return a WP_Error object and `add_action( 'user_profile_update_errors', 'Airstory\Settings\profile_error_save_token' );` (which will show the error message created in 04891d4). Since `Airstory\Settings\profile_error_save_token()` is practically identical to `Airstory\Connection\\user_connection_error()`, remove the latter since the error is specific to the user profile screen.
If WP_Error were throwable, we'd be able to wrap the `do_action()` call at the end of `save_profile_settings()` in a try/catch block and eliminate the multiple calls to `add_action( 'user_profile_update_errors' )`, but that isn't currently the case. At some point in the future, it may be worth refactoring error handling within the connection.