Only require Mcrypt for sites running PHP < 7.0
The Mcrypt requirement was born out of the need for mcrypt_create_iv(), which serves as the fallback for sites that don't have access to the random_bytes() function (introduced in PHP 7.0).

As such, sites running PHP 7.0 or newer don't need Mcrypt and, the library itself was deprecated in 7.1 and removed in 7.2 (http://php.net/manual/en/intro.mcrypt.php).

Fixes #81.